+++ rector.php
@@ -5,10 +5,9 @@ declare(strict_types=1);
 use Rector\Config\RectorConfig;
 use Rector\Doctrine\Set\DoctrineSetList;
 use Rector\Naming\Rector\Class_\RenamePropertyToMatchTypeRector;
-use Rector\Nette\Set\NetteSetList;
-use Rector\Php80\Rector\FunctionLike\UnionTypesRector;
 use Rector\Set\ValueObject\LevelSetList;
 use Rector\Set\ValueObject\SetList;
+use Rector\TypeDeclaration\Rector\ClassMethod\ReturnUnionTypeRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->paths([__DIR__ . '/src', __DIR__ . '/tests', __DIR__ . '/utils']);
@@ -18,7 +17,7 @@ return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->skip([
         RenamePropertyToMatchTypeRector::class => [__DIR__ . '/tests/ORM/'],
 
-        UnionTypesRector::class => [
+        ReturnUnionTypeRector::class => [
             // to keep BC return types
             __DIR__ . '/src/Contract/Entity',
             'src/Model/*/*Trait.php',
@@ -31,7 +30,6 @@ return static function (RectorConfig $rectorConfig): void {
         SetList::DEAD_CODE,
         SetList::CODE_QUALITY,
         SetList::CODING_STYLE,
-        NetteSetList::NETTE_CODE_QUALITY,
         SetList::NAMING,
         LevelSetList::UP_TO_PHP_80,
     ]);