+++ src/Bundle/DependencyInjection/DoctrineBehaviorsExtension.php
@@ -11,10 +11,7 @@ use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
 
 final class DoctrineBehaviorsExtension extends Extension
 {
-    /**
-     * @param string[] $configs
-     */
-    public function load(array $configs, ContainerBuilder $containerBuilder): void
+    public function load(array $configs, ContainerBuilder $containerBuilder)
     {
         $phpFileLoader = new PhpFileLoader($containerBuilder, new FileLocator(__DIR__ . '/../../../config'));
         $phpFileLoader->load('services.php');