+++ CHANGELOG.md
@@ -13,7 +13,7 @@ PRs and issues are linked, so you can find more about it. Thanks to [ChangelogLi
 
 ### Changed
 
-- Update for newer versions of symfony, php and doctrine after fork original KnpLabs/DoctrineBehaviors [@schvoy]
+- Update for newer versions of symfony, php, doctrine and etc. after fork original KnpLabs/DoctrineBehaviors [@schvoy]
 - Remove deprecations [@schvoy]
 
 ## [v2.0.1] - 2020-01-08